+++ packages/bignumber/lib.esm/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "bignumber/5.0.11";
+export declare const version = "bignumber/5.0.12";