+++ packages/bytes/lib.esm/index.js
@@ -34,7 +34,7 @@ export function isBytes(value) {
     }
     for (let i = 0; i < value.length; i++) {
         const v = value[i];
-        if (v < 0 || v >= 256 || (v % 1)) {
+        if (typeof (v) !== "number" || v < 0 || v >= 256 || (v % 1)) {
             return false;
         }
     }