+++ packages/tests/lib.esm/test-utils.js
@@ -509,6 +509,16 @@ describe("BigNumber", function () {
             assert.ok(value.abs().eq(expected));
         });
     });
+    // Fails to create from BN (or any junk with a length) (See: #1172)
+    it("Fails on junk with a length property", function () {
+        const junk = { negative: 0, words: [1000], length: 1, red: null };
+        assert.throws(() => {
+            const value = ethers.BigNumber.from("100").add(junk);
+            console.log("ERROR", value);
+        }, (error) => {
+            return true;
+        });
+    });
     // @TODO: Add more tests here
 });
 describe("FixedNumber", function () {